tests/wpt/tests/pointerevents/compat/pointerevent_mouse_target_after_pointer_target_slot_removed.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
<title>Retarget source mouse event target to slot parent if the target slot of preceding pointer event is removed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<template id="template">
  <div id="slotParent" style="height: 100px">
    <slot id="slot">slot</slot>
  </div>
</template>

<script>
"use strict";

customElements.define(
  "my-elem",
  class extends HTMLElement {
    constructor() {
      super();
      const content = document.getElementById("template").content;
      const shadowRoot = this.attachShadow({ mode: "open" });
      shadowRoot.appendChild(content.cloneNode(true));
    }
  },
);

addEventListener("load", () => {
  async function doTest(t, sourceEventType, precedingEventType) {
    const host = document.createElement("my-elem");
    host.setAttribute("id", "host");
    host.setAttribute("style", "height: 100px");
    const child = document.createElement("div");
    child.setAttribute("id", "child");
    child.setAttribute("style", "height: 100px");
    host.appendChild(child);
    const parent = document.getElementById("parent");
    parent.appendChild(host);
    t.add_cleanup(() => {
      host.remove();
    });
    const slot = host.shadowRoot.getElementById("slotParent").firstElementChild;

    const events = [];
    function logEvent(event) {
      if (!events.length && event.type == sourceEventType) {
        return; // Ignore mouse events before removing the shadow DOM.
      }
      events.push(`${event.type}@${event.target.id ? event.target.id : event.target.localName}`);
    }
    host.shadowRoot.addEventListener(sourceEventType, logEvent, {capture: true});
    child.addEventListener(precedingEventType, event => {
      logEvent(event);
      slot.remove();
      events.push("(slot-removed)");
    });
    const childRect = child.getBoundingClientRect();
    if (sourceEventType == "mousemove") {
      await new test_driver.Actions()
        .pointerMove(0, 0, {origin: document.getElementById("initialPosition")})
        .pointerMove(
          childRect.x + childRect.width / 2,
          childRect.y + childRect.height / 2,
          {origin: "viewport"}
        )
        .send();
    } else {
      await new test_driver.Actions()
        .pointerMove(0, 0, {origin: document.getElementById("initialPosition")})
        .pointerMove(
          childRect.x + childRect.width / 2,
          childRect.y + childRect.height / 2,
          {origin: "viewport"}
        )
        .pointerDown()
        .pointerUp()
        .send();
    }
    removeEventListener(sourceEventType, logEvent, {capture: true});
    assert_equals(
      events.join(", "),
      [
        `${precedingEventType}@child`,
        "(slot-removed)",
        `${sourceEventType}@slotParent`,
      ].join(", ")
    );
  }
  promise_test(async t => {
    await doTest(t, "mousemove", "pointermove");
  }, "mousemove event should be fired on the slot parent when pointermove event listener removes the slot");
  promise_test(async t => {
    await doTest(t, "mousedown", "pointerdown");
  }, "mousedown event should be fired on the slot parent when pointerdown event listener removes the slot");
  promise_test(async t => {
    await doTest(t, "mouseup", "pointerup");
  }, "mouseup event should be fired on the slot parent when pointerup event listener removes the slot");
}, {once: true});
</script>
<body>
  <div id="initialPosition">Initial Position</div>
  <div id="parent" style="height: 100px"></div>
</body>
</html>
